# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "DuckyAPI",
    "description": "A customer facing api for WildDuck",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [{ "url": "http://localhost:3000" }],
  "components": {
    "securitySchemes": {
      "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" }
    },
    "schemas": {
...
      "DnsTxt": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "ducky._domainkey.example.com",
            "description": "Domain name to which the TXT record should be added"
          },
          "value": {
            "type": "string",
            "example": "v=DKIM1;t=s;p=MIGfMA0...",
            "description": "Value of the TXT record"
          }
        },
        "required": ["name", "value"]
      },
      "DkimKey": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "59ef21aef255ed1d9d790e7a",
            "description": "Unique id of the DKIM key"
          },
          "domain": {
            "type": "string",
            "example": "example.com",
            "description": "The domain this DKIM key applies to"
          },
          "selector": {
            "type": "string",
            "example": "ducky",
            "description": "DKIM selector"
          },
          "fingerprint": {
            "type": "string",
            "example": "6a:aa:d7:ba:e4:99:b4:12:e0:f3:35:01:71:d4:f1:d6:b4:95:c4:f5",
            "description": "Unique id of the DKIM key"
          },
          "publicKey": {
            "type": "string",
            "example": "-----BEGIN PUBLIC KEY-----\r\nMIGfMA0...",
            "description": "Public key in DNS format (no prefix/suffix, single line)"
          },
          "dnsTxt": {
            "allOf": [
              { "$ref": "#/components/schemas/DnsTxt" },
              { "description": "Value for the DNS TXT record" }
            ]
          },
          "created": {
            "type": "string",
            "example": "2017-10-24T11:19:10.911Z",
            "description": "Datestring of the time the DKIM key was created"
          }
        },
        "required": [
          "id",
          "domain",
          "selector",
          "fingerprint",
          "publicKey",
          "dnsTxt",
          "created"
        ]
      },
      "AddDkimDto": {
        "type": "object",
        "properties": {
          "selector": {
            "type": "string",
            "example": "default",
            "description": "Selector for dkim key"
          },
          "privateKey": {
            "type": "string",
            "example": "-----BEGIN RSA PRIVATE KEY-----...",
            "description": "Pem formatted DKIM private key. If not set then a new 2048 bit RSA key is generated, beware though that it can take several seconds to complete"
          }
        },
        "required": ["selector"]
      },
...
    }
  },
  "paths": {
...
    "/domains/{domainOrAlias}/dkim": {
      "delete": {
        "operationId": "deleteDkim",
        "summary": "Delete DKIM key for a domain",
        "parameters": [
          {
            "name": "domainOrAlias",
            "required": true,
            "in": "path",
            "description": "example.com",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "DKIM key successfully deleted" },
          "400": { "description": "Bad user input" },
          "401": { "description": "Invalid or expired token" },
          "404": { "description": "Domain not found in account" }
        },
        "tags": ["Dkim"],
        "security": [{ "bearer": [] }],
        "x-code-samples": [
          {
            "lang": "Shell",
            "label": "Curl",
            "source": "curl --request DELETE \\\n  --url http://localhost:3000/domains/%7BdomainOrAlias%7D/dkim \\\n  --header 'authorization: Bearer REPLACE_BEARER_TOKEN'"
          },
          {
            "lang": "JavaScript",
            "label": "Node.js",
            "source": "var http = require(\"http\");\n\nvar options = {\n  \"method\": \"DELETE\",\n  \"hostname\": \"localhost\",\n  \"port\": \"3000\",\n  \"path\": \"/domains/%7BdomainOrAlias%7D/dkim\",\n  \"headers\": {\n    \"authorization\": \"Bearer REPLACE_BEARER_TOKEN\"\n  }\n};\n\nvar req = http.request(options, function (res) {\n  var chunks = [];\n\n  res.on(\"data\", function (chunk) {\n    chunks.push(chunk);\n  });\n\n  res.on(\"end\", function () {\n    var body = Buffer.concat(chunks);\n    console.log(body.toString());\n  });\n});\n\nreq.end();"
          },
          {
            "lang": "JavaScript",
            "label": "JavaScript",
            "source": "var data = null;\n\nvar xhr = new XMLHttpRequest();\nxhr.withCredentials = true;\n\nxhr.addEventListener(\"readystatechange\", function () {\n  if (this.readyState === this.DONE) {\n    console.log(this.responseText);\n  }\n});\n\nxhr.open(\"DELETE\", \"http://localhost:3000/domains/%7BdomainOrAlias%7D/dkim\");\nxhr.setRequestHeader(\"authorization\", \"Bearer REPLACE_BEARER_TOKEN\");\n\nxhr.send(data);"
          },
          {
            "lang": "Python",
            "label": "Python3",
            "source": "import http.client\n\nconn = http.client.HTTPConnection(\"localhost:3000\")\n\nheaders = { 'authorization': \"Bearer REPLACE_BEARER_TOKEN\" }\n\nconn.request(\"DELETE\", \"/domains/%7BdomainOrAlias%7D/dkim\", headers=headers)\n\nres = conn.getresponse()\ndata = res.read()\n\nprint(data.decode(\"utf-8\"))"
          },
          {
            "lang": "PHP",
            "label": "PHP",
            "source": "<?php\n\n$curl = curl_init();\n\ncurl_setopt_array($curl, array(\n  CURLOPT_PORT => \"3000\",\n  CURLOPT_URL => \"http://localhost:3000/domains/%7BdomainOrAlias%7D/dkim\",\n  CURLOPT_RETURNTRANSFER => true,\n  CURLOPT_ENCODING => \"\",\n  CURLOPT_MAXREDIRS => 10,\n  CURLOPT_TIMEOUT => 30,\n  CURLOPT_HTTP_VERSION => CURL_HTTP_VERSION_1_1,\n  CURLOPT_CUSTOMREQUEST => \"DELETE\",\n  CURLOPT_HTTPHEADER => array(\n    \"authorization: Bearer REPLACE_BEARER_TOKEN\"\n  ),\n));\n\n$response = curl_exec($curl);\n$err = curl_error($curl);\n\ncurl_close($curl);\n\nif ($err) {\n  echo \"cURL Error #:\" . $err;\n} else {\n  echo $response;\n}"
          },
          {
            "lang": "Java",
            "label": "Java",
            "source": "HttpResponse<String> response = Unirest.delete(\"http://localhost:3000/domains/%7BdomainOrAlias%7D/dkim\")\n  .header(\"authorization\", \"Bearer REPLACE_BEARER_TOKEN\")\n  .asString();"
          },
          {
            "lang": "C#",
            "label": "C#",
            "source": "var client = new RestClient(\"http://localhost:3000/domains/%7BdomainOrAlias%7D/dkim\");\nvar request = new RestRequest(Method.DELETE);\nrequest.AddHeader(\"authorization\", \"Bearer REPLACE_BEARER_TOKEN\");\nIRestResponse response = client.Execute(request);"
          },
          {
            "lang": "C",
            "label": "C",
            "source": "CURL *hnd = curl_easy_init();\n\ncurl_easy_setopt(hnd, CURLOPT_CUSTOMREQUEST, \"DELETE\");\ncurl_easy_setopt(hnd, CURLOPT_URL, \"http://localhost:3000/domains/%7BdomainOrAlias%7D/dkim\");\n\nstruct curl_slist *headers = NULL;\nheaders = curl_slist_append(headers, \"authorization: Bearer REPLACE_BEARER_TOKEN\");\ncurl_easy_setopt(hnd, CURLOPT_HTTPHEADER, headers);\n\nCURLcode ret = curl_easy_perform(hnd);"
          }
        ]
      },
      "get": {
        "operationId": "getDkim",
        "summary": "Get DKIM key info for a domain",
        "parameters": [
          {
            "name": "domainOrAlias",
            "required": true,
            "in": "path",
            "description": "example.com",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "DKIM key info",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DkimKey" }
              }
            }
          },
          "400": { "description": "Bad user input" },
          "401": { "description": "Invalid or expired token" },
          "404": { "description": "Domain not found in account" }
        },
        "tags": ["Dkim"],
        "security": [{ "bearer": [] }],
        "x-code-samples": [
          {
            "lang": "Shell",
            "label": "Curl",
            "source": "curl --request GET \\\n  --url http://localhost:3000/domains/%7BdomainOrAlias%7D/dkim \\\n  --header 'authorization: Bearer REPLACE_BEARER_TOKEN'"
          },
          {
            "lang": "JavaScript",
            "label": "Node.js",
            "source": "var http = require(\"http\");\n\nvar options = {\n  \"method\": \"GET\",\n  \"hostname\": \"localhost\",\n  \"port\": \"3000\",\n  \"path\": \"/domains/%7BdomainOrAlias%7D/dkim\",\n  \"headers\": {\n    \"authorization\": \"Bearer REPLACE_BEARER_TOKEN\"\n  }\n};\n\nvar req = http.request(options, function (res) {\n  var chunks = [];\n\n  res.on(\"data\", function (chunk) {\n    chunks.push(chunk);\n  });\n\n  res.on(\"end\", function () {\n    var body = Buffer.concat(chunks);\n    console.log(body.toString());\n  });\n});\n\nreq.end();"
          },
          {
            "lang": "JavaScript",
            "label": "JavaScript",
            "source": "var data = null;\n\nvar xhr = new XMLHttpRequest();\nxhr.withCredentials = true;\n\nxhr.addEventListener(\"readystatechange\", function () {\n  if (this.readyState === this.DONE) {\n    console.log(this.responseText);\n  }\n});\n\nxhr.open(\"GET\", \"http://localhost:3000/domains/%7BdomainOrAlias%7D/dkim\");\nxhr.setRequestHeader(\"authorization\", \"Bearer REPLACE_BEARER_TOKEN\");\n\nxhr.send(data);"
          },
          {
            "lang": "Python",
            "label": "Python3",
            "source": "import http.client\n\nconn = http.client.HTTPConnection(\"localhost:3000\")\n\nheaders = { 'authorization': \"Bearer REPLACE_BEARER_TOKEN\" }\n\nconn.request(\"GET\", \"/domains/%7BdomainOrAlias%7D/dkim\", headers=headers)\n\nres = conn.getresponse()\ndata = res.read()\n\nprint(data.decode(\"utf-8\"))"
          },
          {
            "lang": "PHP",
            "label": "PHP",
            "source": "<?php\n\n$curl = curl_init();\n\ncurl_setopt_array($curl, array(\n  CURLOPT_PORT => \"3000\",\n  CURLOPT_URL => \"http://localhost:3000/domains/%7BdomainOrAlias%7D/dkim\",\n  CURLOPT_RETURNTRANSFER => true,\n  CURLOPT_ENCODING => \"\",\n  CURLOPT_MAXREDIRS => 10,\n  CURLOPT_TIMEOUT => 30,\n  CURLOPT_HTTP_VERSION => CURL_HTTP_VERSION_1_1,\n  CURLOPT_CUSTOMREQUEST => \"GET\",\n  CURLOPT_HTTPHEADER => array(\n    \"authorization: Bearer REPLACE_BEARER_TOKEN\"\n  ),\n));\n\n$response = curl_exec($curl);\n$err = curl_error($curl);\n\ncurl_close($curl);\n\nif ($err) {\n  echo \"cURL Error #:\" . $err;\n} else {\n  echo $response;\n}"
          },
          {
            "lang": "Java",
            "label": "Java",
            "source": "HttpResponse<String> response = Unirest.get(\"http://localhost:3000/domains/%7BdomainOrAlias%7D/dkim\")\n  .header(\"authorization\", \"Bearer REPLACE_BEARER_TOKEN\")\n  .asString();"
          },
          {
            "lang": "C#",
            "label": "C#",
            "source": "var client = new RestClient(\"http://localhost:3000/domains/%7BdomainOrAlias%7D/dkim\");\nvar request = new RestRequest(Method.GET);\nrequest.AddHeader(\"authorization\", \"Bearer REPLACE_BEARER_TOKEN\");\nIRestResponse response = client.Execute(request);"
          },
          {
            "lang": "C",
            "label": "C",
            "source": "CURL *hnd = curl_easy_init();\n\ncurl_easy_setopt(hnd, CURLOPT_CUSTOMREQUEST, \"GET\");\ncurl_easy_setopt(hnd, CURLOPT_URL, \"http://localhost:3000/domains/%7BdomainOrAlias%7D/dkim\");\n\nstruct curl_slist *headers = NULL;\nheaders = curl_slist_append(headers, \"authorization: Bearer REPLACE_BEARER_TOKEN\");\ncurl_easy_setopt(hnd, CURLOPT_HTTPHEADER, headers);\n\nCURLcode ret = curl_easy_perform(hnd);"
          }
        ]
      },
      "put": {
        "operationId": "updateDkim",
        "summary": "Add or update DKIM key for a domain",
        "parameters": [
          {
            "name": "domainOrAlias",
            "required": true,
            "in": "path",
            "description": "example.com",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AddDkimDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "DKIM key info",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DkimKey" }
              }
            }
          },
          "400": { "description": "Bad user input" },
          "401": { "description": "Invalid or expired token" },
          "404": { "description": "Domain not found in account" }
        },
        "tags": ["Dkim"],
        "security": [{ "bearer": [] }],
        "x-code-samples": [
          {
            "lang": "Shell",
            "label": "Curl",
            "source": "curl --request PUT \\\n  --url http://localhost:3000/domains/%7BdomainOrAlias%7D/dkim \\\n  --header 'authorization: Bearer REPLACE_BEARER_TOKEN' \\\n  --header 'content-type: application/json' \\\n  --data '{\"selector\":\"default\",\"privateKey\":\"-----BEGIN RSA PRIVATE KEY-----...\"}'"
          },
          {
            "lang": "JavaScript",
            "label": "Node.js",
            "source": "var http = require(\"http\");\n\nvar options = {\n  \"method\": \"PUT\",\n  \"hostname\": \"localhost\",\n  \"port\": \"3000\",\n  \"path\": \"/domains/%7BdomainOrAlias%7D/dkim\",\n  \"headers\": {\n    \"content-type\": \"application/json\",\n    \"authorization\": \"Bearer REPLACE_BEARER_TOKEN\"\n  }\n};\n\nvar req = http.request(options, function (res) {\n  var chunks = [];\n\n  res.on(\"data\", function (chunk) {\n    chunks.push(chunk);\n  });\n\n  res.on(\"end\", function () {\n    var body = Buffer.concat(chunks);\n    console.log(body.toString());\n  });\n});\n\nreq.write(JSON.stringify({selector: 'default', privateKey: '-----BEGIN RSA PRIVATE KEY-----...'}));\nreq.end();"
          },
          {
            "lang": "JavaScript",
            "label": "JavaScript",
            "source": "var data = JSON.stringify({\n  \"selector\": \"default\",\n  \"privateKey\": \"-----BEGIN RSA PRIVATE KEY-----...\"\n});\n\nvar xhr = new XMLHttpRequest();\nxhr.withCredentials = true;\n\nxhr.addEventListener(\"readystatechange\", function () {\n  if (this.readyState === this.DONE) {\n    console.log(this.responseText);\n  }\n});\n\nxhr.open(\"PUT\", \"http://localhost:3000/domains/%7BdomainOrAlias%7D/dkim\");\nxhr.setRequestHeader(\"content-type\", \"application/json\");\nxhr.setRequestHeader(\"authorization\", \"Bearer REPLACE_BEARER_TOKEN\");\n\nxhr.send(data);"
          },
          {
            "lang": "Python",
            "label": "Python3",
            "source": "import http.client\n\nconn = http.client.HTTPConnection(\"localhost:3000\")\n\npayload = \"{\\\"selector\\\":\\\"default\\\",\\\"privateKey\\\":\\\"-----BEGIN RSA PRIVATE KEY-----...\\\"}\"\n\nheaders = {\n    'content-type': \"application/json\",\n    'authorization': \"Bearer REPLACE_BEARER_TOKEN\"\n    }\n\nconn.request(\"PUT\", \"/domains/%7BdomainOrAlias%7D/dkim\", payload, headers)\n\nres = conn.getresponse()\ndata = res.read()\n\nprint(data.decode(\"utf-8\"))"
          },
          {
            "lang": "PHP",
            "label": "PHP",
            "source": "<?php\n\n$curl = curl_init();\n\ncurl_setopt_array($curl, array(\n  CURLOPT_PORT => \"3000\",\n  CURLOPT_URL => \"http://localhost:3000/domains/%7BdomainOrAlias%7D/dkim\",\n  CURLOPT_RETURNTRANSFER => true,\n  CURLOPT_ENCODING => \"\",\n  CURLOPT_MAXREDIRS => 10,\n  CURLOPT_TIMEOUT => 30,\n  CURLOPT_HTTP_VERSION => CURL_HTTP_VERSION_1_1,\n  CURLOPT_CUSTOMREQUEST => \"PUT\",\n  CURLOPT_POSTFIELDS => \"{\\\"selector\\\":\\\"default\\\",\\\"privateKey\\\":\\\"-----BEGIN RSA PRIVATE KEY-----...\\\"}\",\n  CURLOPT_HTTPHEADER => array(\n    \"authorization: Bearer REPLACE_BEARER_TOKEN\",\n    \"content-type: application/json\"\n  ),\n));\n\n$response = curl_exec($curl);\n$err = curl_error($curl);\n\ncurl_close($curl);\n\nif ($err) {\n  echo \"cURL Error #:\" . $err;\n} else {\n  echo $response;\n}"
          },
          {
            "lang": "Java",
            "label": "Java",
            "source": "HttpResponse<String> response = Unirest.put(\"http://localhost:3000/domains/%7BdomainOrAlias%7D/dkim\")\n  .header(\"content-type\", \"application/json\")\n  .header(\"authorization\", \"Bearer REPLACE_BEARER_TOKEN\")\n  .body(\"{\\\"selector\\\":\\\"default\\\",\\\"privateKey\\\":\\\"-----BEGIN RSA PRIVATE KEY-----...\\\"}\")\n  .asString();"
          },
          {
            "lang": "C#",
            "label": "C#",
            "source": "var client = new RestClient(\"http://localhost:3000/domains/%7BdomainOrAlias%7D/dkim\");\nvar request = new RestRequest(Method.PUT);\nrequest.AddHeader(\"content-type\", \"application/json\");\nrequest.AddHeader(\"authorization\", \"Bearer REPLACE_BEARER_TOKEN\");\nrequest.AddParameter(\"application/json\", \"{\\\"selector\\\":\\\"default\\\",\\\"privateKey\\\":\\\"-----BEGIN RSA PRIVATE KEY-----...\\\"}\", ParameterType.RequestBody);\nIRestResponse response = client.Execute(request);"
          },
          {
            "lang": "C",
            "label": "C",
            "source": "CURL *hnd = curl_easy_init();\n\ncurl_easy_setopt(hnd, CURLOPT_CUSTOMREQUEST, \"PUT\");\ncurl_easy_setopt(hnd, CURLOPT_URL, \"http://localhost:3000/domains/%7BdomainOrAlias%7D/dkim\");\n\nstruct curl_slist *headers = NULL;\nheaders = curl_slist_append(headers, \"content-type: application/json\");\nheaders = curl_slist_append(headers, \"authorization: Bearer REPLACE_BEARER_TOKEN\");\ncurl_easy_setopt(hnd, CURLOPT_HTTPHEADER, headers);\n\ncurl_easy_setopt(hnd, CURLOPT_POSTFIELDS, \"{\\\"selector\\\":\\\"default\\\",\\\"privateKey\\\":\\\"-----BEGIN RSA PRIVATE KEY-----...\\\"}\");\n\nCURLcode ret = curl_easy_perform(hnd);"
          }
        ]
      }
    },
...
  }
}